+++ src/evalBlock.js
@@ -2,6 +2,10 @@ class EvalBlock {
   constructor(string) {
     this.string = string;
   }
+
+  toString() {
+    return `EvalBlock('${this.string}')`;
+  }
 }
 
 exports.EvalBlock = EvalBlock;