+++ src/mode.js
@@ -3,6 +3,10 @@ class Mode {
     this.name = name;
     this.rules = [];
   }
+
+  toString() {
+    return `Mode{name: ${this.name}, rules: ${this.rules}}`;
+  }
 }
 
 exports.Mode = Mode;