+++ CHANGELOG.md
@@ -4,3 +4,10 @@
   marker
 * Fixed a bug with backslash escapes inside literal blocks so literal double square
   brackets can be used [[ inside blocks like this \\]] ]]
+
+### 0.0.3
+* Add WeightedChoice class and use it in place of weight objects.
+  `{option: ___, weight: ___} -> new WeightedChoice(choice, weight)`
+  The new class has a `choice` property in place of an `option` one.
+* rename `rand.weightedChoice() -> rand.weightedChoose()`
+* Implement toString() methods in all classes